+++ packages/picasso/src/test-utils/index.tsx
@@ -29,3 +29,4 @@ const customRender = (
 
 export * from '@testing-library/react'
 export { customRender as render, TestingPicasso }
+export * from './mock-data'
