+++ packages/picasso/src/test-utils/mock-data/index.ts
@@ -1,0 +1,1 @@
+export { default as tableMockData } from './table'
